remove need to use forking while still supporting timeouts
