+++ src/Illuminate/Queue/Console/ListenCommand.php
@@ -2,7 +2,6 @@
 
 namespace Illuminate\Queue\Console;
 
-use RuntimeException;
 use Illuminate\Queue\Listener;
 use Illuminate\Console\Command;
 use Symfony\Component\Console\Input\InputOption;
@@ -48,8 +47,6 @@ class ListenCommand extends Command
      * Execute the console command.
      *
      * @return void
-     *
-     * @throws \RuntimeException
      */
     public function fire()
     {
@@ -65,10 +62,6 @@ class ListenCommand extends Command
         $connection = $this->input->getArgument('connection');
 
         $timeout = $this->input->getOption('timeout');
-
-        if ($timeout && ! function_exists('pcntl_fork')) {
-            throw new RuntimeException('Timeouts not supported without the pcntl_fork extension.');
-        }
 
         // We need to get the right queue for the connection which is set in the queue
         // configuration file for the application. We will pull it based on the set
