+++ src/Illuminate/Queue/Console/WorkCommand.php
@@ -3,13 +3,8 @@
 namespace Illuminate\Queue\Console;
 
 use Carbon\Carbon;
-use RuntimeException;
-use Illuminate\Queue\Worker;
 use Illuminate\Console\Command;
-use Illuminate\Queue\WorkerOptions;
-use Illuminate\Contracts\Queue\Job;
-use Illuminate\Queue\Events\JobFailed;
-use Illuminate\Queue\Events\JobProcessed;
+use Symfony\Component\Process\Process;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Input\InputArgument;
 
@@ -27,27 +22,7 @@ class WorkCommand extends Command
      *
      * @var string
      */
-    protected $description = 'Process the next job on a queue';
-
-    /**
-     * The queue worker instance.
-     *
-     * @var \Illuminate\Queue\Worker
-     */
-    protected $worker;
-
-    /**
-     * Create a new queue listen command.
-     *
-     * @param  \Illuminate\Queue\Worker  $worker
-     * @return void
-     */
-    public function __construct(Worker $worker)
-    {
-        parent::__construct();
-
-        $this->worker = $worker;
-    }
+    protected $description = 'Start processing jobs from the queue';
 
     /**
      * Execute the console command.
@@ -56,136 +31,27 @@ class WorkCommand extends Command
      */
     public function fire()
     {
-        if ($this->downForMaintenance() && $this->option('once')) {
-            return $this->worker->sleep($this->option('sleep'));
-        }
+        $process = $this->newProxyProcess();
 
-        // We'll listen to the processed and failed events so we can write information
-        // to the console as jobs are processed, which will let the developer watch
-        // which jobs are coming through a queue and be informed on its progress.
-        $this->listenForEvents();
-
-        $connection = $this->argument('connection')
-                        ?: $this->laravel['config']['queue.default'];
-
-        // We need to get the right queue for the connection which is set in the queue
-        // configuration file for the application. We will pull it based on the set
-        // connection being run for the queue operation currently being executed.
-        $queue = $this->option('queue') ?: $this->laravel['config']->get(
-            "queue.connections.{$connection}.queue", 'default'
-        );
-
-        $response = $this->runWorker(
-            $connection, $queue
-        );
+        exit($process->run(function ($type, $line) {
+            if (trim($line) !== '.') {
+                $this->output->write($line);
+            }
+        }));
     }
 
     /**
-     * Run the worker instance.
+     * Get a new proxy process to the daemon command.
      *
-     * @param  string  $connection
-     * @param  string  $queue
-     * @return array
+     * @return Process
      */
-    protected function runWorker($connection, $queue)
+    protected function newProxyProcess()
     {
-        $this->worker->setCache($this->laravel['cache']->driver());
+        $_SERVER['argv'][1] = 'queue:daemon';
 
-        $method = $this->option('once') ? 'runNextJob' : 'daemon';
-
-        return $this->worker->{$method}(
-            $connection, $queue, $this->gatherWorkerOptions()
-        );
-    }
-
-    /**
-     * Gather all of the queue worker options as a single object.
-     *
-     * @return \Illuminate\Queue\WorkerOptions
-     */
-    protected function gatherWorkerOptions()
-    {
-        $timeout = $this->option('timeout');
-
-        if ($timeout && ! function_exists('pcntl_fork')) {
-            throw new RuntimeException('The pcntl extension is required in order to specify job timeouts.');
-        }
-
-        return new WorkerOptions(
-            $this->option('delay'), $this->option('memory'),
-            $timeout, $this->option('sleep'),
-            $this->option('tries')
-        );
-    }
-
-    /**
-     * Listen for the queue events in order to update the console output.
-     *
-     * @return void
-     */
-    protected function listenForEvents()
-    {
-        $this->laravel['events']->listen(JobProcessed::class, function ($event) {
-            $this->writeOutput($event->job, false);
-        });
-
-        $this->laravel['events']->listen(JobFailed::class, function ($event) {
-            $this->writeOutput($event->job, true);
-
-            $this->logFailedJob($event);
-        });
-    }
-
-    /**
-     * Write the status output for the queue worker.
-     *
-     * @param  \Illuminate\Contracts\Queue\Job  $job
-     * @param  bool  $failed
-     * @return void
-     */
-    protected function writeOutput(Job $job, $failed)
-    {
-        if ($failed) {
-            $this->output->writeln('<error>['.Carbon::now()->format('Y-m-d H:i:s').'] Failed:</error> '.$job->resolveName());
-        } else {
-            $this->output->writeln('<info>['.Carbon::now()->format('Y-m-d H:i:s').'] Processed:</info> '.$job->resolveName());
-        }
-    }
-
-    /**
-     * Store a failed job event.
-     *
-     * @param  JobFailed  $event
-     * @return void
-     */
-    protected function logFailedJob(JobFailed $event)
-    {
-        $this->laravel['queue.failer']->log(
-            $event->connectionName, $event->job->getQueue(),
-            $event->job->getRawBody(), $event->exception
-        );
-    }
-
-    /**
-     * Determine if the worker should run in maintenance mode.
-     *
-     * @return bool
-     */
-    protected function downForMaintenance()
-    {
-        return $this->option('force') ? false : $this->laravel->isDownForMaintenance();
-    }
-
-    /**
-     * Get the console command arguments.
-     *
-     * @return array
-     */
-    protected function getArguments()
-    {
-        return [
-            ['connection', InputArgument::OPTIONAL, 'The name of connection', null],
-        ];
+        return (new Process(PHP_BINARY.' '.implode(' ', $_SERVER['argv']), getcwd()))
+                    ->setTimeout(null)
+                    ->setIdleTimeout($this->option('timeout') + $this->option('sleep'));
     }
 
     /**
